Added "Today I Learned" (TIL) to each file
These serve as a way to track what I'm learning about the languages,
gotchas, snippets, etc.

Pros:
++ SO FAST
++ easy to test, write, and compile
+ structs over classes is simpler and typedefs are slick
+ pointer auto-handling is convienent
+ good built-in string functions
+ built-in slices and maps are convienent

Cons:
- capital/lowercase for exports is really goofy
- index, element for slice `range` is really dumb
    - really should be element, index but that would make it backward
    from how `range` works with maps.
    - either way, an `elements` keyword or `values` would've been great
- slices aren't always intuitive (still haven't fully groked them)
    - see cloning of slices
- no negative index handling on slices (like python)
- no ternary operator or one-line if-statements makes code overly
verbose
- regexs w/o backreferences is super limited
- lack of generics means writing a lot of low level functions by hand,
like `contains/indexOf` and such
- no function overloading

In general, doing these problems in Go has been really enjoyable so far.

It's not quite python but the speed is unbelievable and the compiler
checks make a strong case for Go over Python.

The move away from OOP is a fresh experience. When I want to do simple
things, I don't want to start with `public class Yada....`.

The simplicity of the language means that I *could* write this all in vi
though the vim plugins make it a lot better.